+++ example/lib/main.dart
@@ -1,8 +1,8 @@
+import 'package:bottom_bar/pages/profile.dart';
 import 'package:flutter/material.dart';
 import 'package:instagram_bottom_nav_bar/instagram_tab_view.dart';
-import 'package:instagram_bottom_nav_bar/widget/instagram_tab_item.dart';
+import 'package:instagram_bottom_nav_bar/widget/icon_type.dart';
 import 'pages/home_page.dart';
-import 'pages/messages.dart';
 
 main() {
   runApp(const MyApp());
@@ -18,15 +18,37 @@ class MyApp extends StatelessWidget {
       home: InstagramTabView(
         selectedItemColor: Colors.black,
         unselectedItemColor: Colors.grey.shade400,
-        backgroundColor: Colors.white, // Default is Colors.white
-        elevation: 12, // Default elevation is 12,
-        showSelectedLabels: true, // Default showSelectedLabels is true,
-        showUnselectedLabels: true, // Default showUnselectedLabels is true
-        iconSize: 24, // Default iconSize is 24,
+        backgroundColor: Colors.white,
+        // Default is Colors.white
+        elevation: 12,
+        // Default elevation is 12,
+        showSelectedLabels: true,
+        // Default showSelectedLabels is true,
+        showUnselectedLabels: true,
+        // Default showUnselectedLabels is true
+        iconSize: 24,
+        // Default iconSize is 24,
         bottomNavigationBarType: BottomNavigationBarType.fixed,
+        iconType: IconType.icon,
         items: [
-          InstagramTabItem(label: 'Home Page', page: HomePage(), icon: Icon(Icons.add)),
-          InstagramTabItem(label: 'Profile', page: Messages(), icon: Icon(Icons.refresh)),
+          InstagramTabItem(
+            label: 'Home Page',
+            page: HomePage(),
+            icon: Icons.add,
+          ),
+          InstagramTabItem(
+            label: 'Home Page',
+            page: HomePage(),
+            icon: Icons.add,
+            settings: (settings) {
+              switch (settings.name) {
+                default:
+                  return MaterialPageRoute(
+                    builder: (context) => Profile(),
+                  );
+              }
+            },
+          )
         ],
       ),
     );