+++ lib/instagram_tab_view.dart
@@ -2,8 +2,12 @@ library instagram_bottom_nav_bar;
 
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
+import 'package:flutter_svg/flutter_svg.dart';
+import 'package:instagram_bottom_nav_bar/widget/icon_type.dart';
 import 'widget/instagram_tab_item.dart';
+export 'widget/instagram_tab_item.dart';
 import 'widget/proste_indexed_stack.dart';
 
 part 'view/bottom_nav_bar_view.dart';
 part 'const/bottom_nav_bar_const.dart';
+