+++ lib/widget/icon_type.dart
@@ -1,0 +1,1 @@
+enum IconType { icon, pngOrJpg, svg }