+++ lib/widget/instagram_tab_item.dart
@@ -3,8 +3,15 @@ import 'package:flutter/material.dart';
 class InstagramTabItem {
   String label;
   Widget page;
-  Widget icon;
+  dynamic icon;
+  String? initialRoute;
+  Route<dynamic>? Function(RouteSettings)? settings;
 
-  InstagramTabItem(
-      {required this.label, required this.page, required this.icon});
+  InstagramTabItem({
+    required this.label,
+    required this.page,
+    required this.icon,
+    this.initialRoute,
+    this.settings
+  });
 }